+++ src/store/sagas/index.ts
@@ -1,8 +1,9 @@
 import { takeEvery } from 'redux-saga/effects';
 import { ActionTypes } from '../actions';
-import { logoutSaga, checkAuthTimeoutSaga } from './auth';
+import { logoutSaga, checkAuthTimeoutSaga, authUserSaga } from './auth';
 
 export function* rootSaga() {
   yield takeEvery(ActionTypes.AUTH_INITIATE_LOGOUT, logoutSaga);
   yield takeEvery(ActionTypes.AUTH_CHECK_TIMEOUT, checkAuthTimeoutSaga);
+  yield takeEvery(ActionTypes.AUTH_USER, authUserSaga);
 }
